+++ SIL.Windows.Forms.Archiving/IMDI/IMDIArchivingDlg.cs
@@ -345,7 +345,6 @@ namespace SIL.Windows.Forms.Archiving.IMDI
 			//UpdateLaunchButtonText();
 			_buttonLaunchRamp.Visible = false;
 			_tableLayoutPanel.SetColumn(_buttonCreatePackage, 1);
-			_buttonCancel.Text = LocalizationManager.GetString("DialogBoxes.IMDIArchivingDlg.CloseButtonLabel", "Close");
 			_buttonCreatePackage.Text = LocalizationManager.GetString("DialogBoxes.IMDIArchivingDlg.CreatePackageButtonLabel", "Create Package");
 			UpdateOverviewText();
 		}